# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Habitat Commissioner",
    "description": "Express REST API for moving data between orgs and environments in apigee",
    "license": {
      "name": "MIT",
      "url": "http://github.com/gruntjs/grunt/blob/master/LICENSE-MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/notification-email-templates/apigee/list": {
      "get": {
        "description": "List of email templates from the specified org in Apigee",
        "summary": "List email templates",
        "tags": [
          "notification email templates"
        ],
        "operationId": "listEmailTemplates",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/notification-email-templates/apigee/details/{template_id}": {
      "get": {
        "description": "Details of a template from Apigee",
        "summary": "Get an email template",
        "tags": [
          "notification email templates"
        ],
        "operationId": "getEmailTemplate",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "email template ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/notification-email-templates/repo/list": {
      "get": {
        "description": "List of templates from the repo",
        "summary": "List email templates from the repo",
        "tags": [
          "notification email templates"
        ],
        "operationId": "listRepoEmailTemplates",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "repo",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "repo name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/notification-email-templates/repo/details/{template_id}": {
      "get": {
        "description": "Details of a template from the specified repo",
        "summary": "Get an email template from the repo",
        "tags": [
          "notification email templates"
        ],
        "operationId": "getRepoEmailTemplate",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "repo",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "repo name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "email template ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/notification-email-templates/repo": {
      "post": {
        "description": "Export email templates from Apigee to the repo",
        "summary": "Export templates",
        "tags": [
          "notification email templates"
        ],
        "operationId": "exportEmailTemplates",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "repo",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "repo name"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ApigeeRequest"
            }
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/notification-email-templates/apigee": {
      "put": {
        "description": "Update an existing email template in Apigee using data from the repo",
        "summary": "Update email templates",
        "tags": [
          "notification email templates"
        ],
        "operationId": "updateEmailTemplates",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "repo",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "repo name"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ApigeeRequest"
            }
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      },
      "post": {
        "description": "Import email templates to Apigee using data from the repo",
        "summary": "Import email templates",
        "tags": [
          "notification email templates"
        ],
        "operationId": "importEmailTemplates",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "repo",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "repo name"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ApigeeRequest"
            }
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ApigeeResponse": {
      "title": "apigeeResponse",
      "type": "object",
      "properties": {
        "code": {
          "description": "",
          "example": 200,
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "description": "",
          "example": "Context sensitive message",
          "type": "string"
        },
        "data": {
          "description": "Generally reflects the response directly from the Apigee management API. May be an object or an array",
          "type": "object"
        }
      }
    },
    "ApigeeRequest": {
      "type": "array",
      "items": {
        "type": "string",
        "description": "item name or ID",
        "example": "item-name-or-id"
      }
    },
...
  }
}